Completes Module for Iteration 0
